+++ src/createContext.ts
@@ -6,11 +6,12 @@ import { Selector, Context } from './interfaces';
  * @param useValue function for init context state
  * @param selectors hooks with context
  */
-export function createContext<Props extends {}, Value extends Record<string, any>, Selectors extends Selector<Value>[]>(
-    useValue: (props: Props) => Value,
-    ...selectors: Selectors
-) {
-    const injectionKey: InjectionKey<Context<Value, Selectors>> = Symbol();
+export function createContext<
+    Props extends {},
+    Value extends Record<string, any>,
+    Selectors extends Selector<Value, Props>[]
+>(useValue: (props: Props) => Value, ...selectors: Selectors) {
+    const injectionKey: InjectionKey<Context<Value, Props, Selectors>> = Symbol();
 
     const NO_PROVIDER = {};
 
@@ -19,13 +20,13 @@ export function createContext<Props extends {}, Value extends Record<string, any
             defineComponent({
                 name: 'Provider',
                 setup() {
-                    const context = useValue(props);
+                    const initialContext = useValue(props);
 
                     const hookContextValues = selectors.reduce((merged, selector) => {
-                        return Object.assign({}, merged, selector.call(null, context));
+                        return Object.assign({}, merged, selector.call(null, initialContext, props));
                     }, Object.create(null));
 
-                    provide(injectionKey, Object.assign({}, context, hookContextValues));
+                    provide(injectionKey, Object.assign({}, initialContext, hookContextValues));
 
                     return () => h(Fragment, slots.default?.());
                 },
@@ -34,7 +35,7 @@ export function createContext<Props extends {}, Value extends Record<string, any
     };
 
     function dispatch() {
-        const context = inject(injectionKey, NO_PROVIDER) as Context<Value, Selectors>;
+        const context = inject(injectionKey, NO_PROVIDER) as Context<Value, Props, Selectors>;
 
         if (context === NO_PROVIDER) {
             console.warn('[vc-state] The ContextProvider is never used.');