+++ src/interfaces.ts
@@ -1,7 +1,14 @@
-export type Selector<Value extends Record<string, any>> = (value: Value) => Record<string, any>;
+export type Selector<Value extends Record<string, any>, Props extends {}> = (
+    initialContext: Value,
+    props: Props
+) => Record<string, any>;
 
-export type DefineContext<Value extends Record<string, any>, Selectors extends Selector<Value>[]> = {
-    [Key in keyof Selectors]: Selectors[Key] extends Selector<Value> ? ReturnType<Selectors[Key]> : {};
+export type DefineContext<
+    Value extends Record<string, any>,
+    Props extends {},
+    Selectors extends Selector<Value, Props>[]
+> = {
+    [Key in keyof Selectors]: Selectors[Key] extends Selector<Value, Props> ? ReturnType<Selectors[Key]> : {};
 };
 
 export type First<F extends Record<string, any>, R extends Record<string, any>[]> = [F, ...R];
@@ -14,5 +21,8 @@ export type MergeContext<H extends Record<string, any>[]> = H extends []
     ? C & MergeContext<R>
     : {};
 
-export type Context<Value extends Record<string, any>, Selectors extends Selector<Value>[]> = Value &
-    MergeContext<DefineContext<Value, Selectors>>;
+export type Context<
+    Value extends Record<string, any>,
+    Props extends {},
+    Selectors extends Selector<Value, Props>[]
+> = Value & MergeContext<DefineContext<Value, Props, Selectors>>;